Teste dia 29/01/2023

+++ testes/talker.cpp
@@ -1,7 +1,7 @@
 #include "ros/ros.h"
 #include "std_msgs/String.h"
 #include <sstream>
-#include "std_msgs/Int32.h"
+#include "std_msgs/Float32.h"
 #include <termios.h>
 
 int main(int argc, char **argv)
@@ -10,13 +10,13 @@ int main(int argc, char **argv)
     ros::init(argc, argv, "talker");
     ros::NodeHandle n;
 
-    ros::Publisher chatter_pub = n.advertise<std_msgs::Int32>("chatter", 10);
+    ros::Publisher chatter_pub = n.advertise<std_msgs::Float32>("chatter", 10);
 
     while (ros::ok())
     {
-        int input;
+        float input;
         std::cin >> input;
-        std_msgs::Int32 msg;
+        std_msgs::Float32 msg;
         msg.data = input;
         chatter_pub.publish(msg);
         ros::spinOnce();
